+++ src/views/NotFound.js
@@ -1,0 +1,12 @@
+import React from 'react'
+import '../App.scss';
+
+const Header = props => {
+    return (
+        <div className="not-found">
+            <h1>Page not found!</h1>
+        </div>
+    )
+}
+
+export default Header