+++ CHANGES.md
@@ -1,5 +1,15 @@
 # Changes
 
+## Change 2
+
+- $ npm install react-router-dom
+- Wrapped App component in <Router> - tags
+- Added <Switch>-Tag in the main-Div
+- Created two folders "components" (for router-independent components) and "views" for (router-dependent components)
+- Added 12 photos, 4 for each hobby in the src/images folder
+- included them in the Hobbies-component
+- in the Header-component, added Router-Links
+
 ## Change 1
 
 Converted functional React-boilerplate component into class-component.